requirements.txt
# PDF處理核心套件
pymupdf==1.25.5
img2table==1.4.1
pytesseract==0.3.13
pillow==11.2.1
opencv-contrib-python==4.11.0.86

# 數據處理
numpy
pandas==2.2.3

# 向量化
torch==2.6.0
transformers==4.51.3
sentence-transformers==2.7.0
FlagEmbedding==1.3.5
huggingface-hub==0.31.1

# LangChain
langchain-community==0.3.23
langchain-core==0.3.59
langchain-huggingface==0.2.0

# 向量資料庫
qdrant-client==1.14.3

# model
boto3==1.40.47

# 其他必要工具
python-dotenv==1.1.0
pydantic==2.11.4
tqdm==4.67.1
jieba==0.42.1

--extra-index-url https://download.pytorch.org/whl/cu124